+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Augmented assignments.
 - Maximum recursion depth to avoid stack overflow.
 - Better variable handling with FxHashMap for performance.
+- `no-module-imports` feature to disable module imports.
+- `no-lib-imports` feature to disable library imports.
+- `i32-ints` feature to use i32 as integers instead of i64.
+- `f32-floats` feature to use f32 as floats instead of f64.
 
 ## [0.1.0] - 2026-05-12